requirements.txt
# requirements file
# Python version: 3.10.13
# Dependencies for the multi-agent conversation project

pyautogen==0.2.25
chess==1.10.0
matplotlib
numpy
pandas
yfinance
pyyaml==6.0.1
backoff==2.2.1
python-dotenv==1.0.1
google-generativeai==0.8.3